Add explicit type casting to avoid compiler error
Seems gcc14 starts to give error instead of warn, which blocks build on
Arch linux. This commit unblocks by adding explicit casting for all
errors.

reference: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=106416

+++ src/legacy/cnxz/nxz_common.c
@@ -9,7 +9,7 @@ uint32_t* sub57DD90(uint32_t* this) {
 
 	v1 = this;
 	v2 = calloc(1, 0x224u);
-	*v1 = v2;
+	*v1 = (uint32_t)v2;
 	memset(v2, 0, 0x224u);
 	return v1;
 }
